feat: Remodeled case in FreeCAD
The previous model was done in OnShape, which is a propritary web-based
platform. It is easy to use, and they give out free accounts for personal
use, however the free accounts could be discontinued at any time, leaving
us with only the .stl files. If that happened, making changes to the model
would be very difficult. By recreating the model in FreeCAD, we don't have to
worry about this.

In addition, this makes the case 4mm taller (protruding from the wall further)
in order to solve some issues we've seen, such as:

- Touchscreen not working because the case is too short for the LCD pins
- Case won't close because the LCD tab collides with the AC/DC converter

This model should match CoverOneSnap2H-taller.stl in every functional way.
There are tiny cosmetic differences.

Finally, it will be easy to make modifications. For example, there is a report
of some models of LCD touchscreens being slightly thicker. With the FreeCAD
model, changing the space for the screen is just a matter of changing the length
of the LCD_tab_posts_pad. Similarly, changing the height again, would just be a
matter of changing a single constraint in the Side_sketch.