Improved documentation and perfected the 35mm edition
The standard case works well for the silent version of the HestiaPi ONE
as long as the correct length LCD headers can be obtained.

However, the clicky edition of the HestiaPi ONE has the temperature
sensor on the back of the board, which makes the board sit higher. This
causes the case to be just a couple millimeters too short. to fit
everything into the case. The 35mm edition of the cover was created to
address this problem.

Also updated the README to help people navigate this repo and added the
height of the original case into the spreadsheet in the FreeCAD model
so people can easily tell if the height has been modified.